Code review.

+++ Gambit/Include/Gambit/Object.hpp
@@ -30,7 +30,7 @@ namespace Gambit
         string Type;
         string Callback;
         Vector2 Position;
-        float Rotation{ 0 };
+        double Rotation{ 0 };
         float Scale{ 1 };
         int Layer{ 0 };
         bool Mirror{ false };
